// src/utils/enums/table.ts
/* eslint-disable no-unused-vars */
export enum TableCellTypes {
  STRING = 'string',
  STRING_WITH_ICON = 'string_with_icon',
  STRING_WITH_SLICE = 'string_with_slice',
  STRING_WITH_CUT = 'string_with_cut',
  DATE = 'date',
  STATUS = 'status',
  AMOUNT = 'amount'
}

export enum SortingDirection {
  ascending = 'ascending',
  descending = 'descending',
  default = 'sort-default'
}